サイドメニューのアイテムを各コンポーネントに分離
